+++ mods/CreateEnchantmentIndustry/src/main/java/plus/dragons/createenchantmentindustry/common/registry/CEIDataMaps.java
@@ -139,7 +139,7 @@ public class CEIDataMaps {
 
     public static void generate(RegistrateDataMapProvider provider) {
         provider.builder(EXPERIENCE_FUEL)
-                .add(CEIItems.EXPERIENCE_BUCKET, ExperienceFuel.normal(1000, Items.BUCKET.getDefaultInstance()), false) // TODO Temporary solution for Create's bug, See https://github.com/Creators-of-Create/Create/pull/8304
+                .add(CEIItems.EXPERIENCE_BUCKET, ExperienceFuel.normal(1000, Items.BUCKET.getDefaultInstance()), false) // Workaround solution, See https://github.com/Creators-of-Create/Create/pull/8304
                 .add(CEIItems.EXPERIENCE_CAKE, ExperienceFuel.special(1000), false)
                 .add(CEIItems.EXPERIENCE_CAKE_SLICE, ExperienceFuel.special(250), false)
                 .add(CEIBlocks.SUPER_EXPERIENCE_BLOCK.getId(), ExperienceFuel.special(27), false)