+++ mods/CreateEnchantmentIndustry/src/main/java/plus/dragons/createenchantmentindustry/config/CEIEnchantmentsConfig.java
@@ -27,6 +27,15 @@ public class CEIEnchantmentsConfig extends ConfigBase {
     public final ConfigInt blazeEnchanterMaxSuperEnchantLevel = i(60, 0,
             "blazeEnchanterMaxSuperEnchantLevel",
             Comments.blazeEnchanterMaxSuperEnchantLevel);
+    public final ConfigFloat blazeEnchanterBlockedLightningCurseChance = f(0.5f, 0.0f, 1.0f,
+            "blazeEnchanterBlockedLightningCurseChance",
+            Comments.blazeEnchanterBlockedLightningCurseChance);
+    public final ConfigInt blazeEnchanterBlockedLightningCurseCount = i(1, 0, 16,
+            "blazeEnchanterBlockedLightningCurseCount",
+            Comments.blazeEnchanterBlockedLightningCurseCount);
+    public final ConfigInt blazeEnchanterBlockedLightningCurseMaxLevel = i(1, 1, 255,
+            "blazeEnchanterBlockedLightningCurseMaxLevel",
+            Comments.blazeEnchanterBlockedLightningCurseMaxLevel);
     public final ConfigInt enchantmentMaxLevelExtension = i(1, 0, 255,
             "enchantmentMaxLevelExtension",
             Comments.enchantmentMaxLevelExtension);
@@ -45,6 +54,9 @@ public class CEIEnchantmentsConfig extends ConfigBase {
     static class Comments {
         static final String blazeEnchanterMaxEnchantLevel = "The max experience level a Blaze Enchanter can use in Regular Enchanting";
         static final String blazeEnchanterMaxSuperEnchantLevel = "The max experience level a Blaze Enchanter can use in Super Enchanting";
+        static final String blazeEnchanterBlockedLightningCurseChance = "Chance per curse roll for blocked-lightning Super Enchanting to add an applicable curse to the result";
+        static final String blazeEnchanterBlockedLightningCurseCount = "Maximum curse rolls for blocked-lightning Super Enchanting";
+        static final String blazeEnchanterBlockedLightningCurseMaxLevel = "Maximum curse level blocked-lightning Super Enchanting can add";
         static final String enchantmentMaxLevelExtension = "Max enchantment level in Super Enchanting will be extended by this value";
         static final String ignoreEnchantmentCompatibility = "If Super Enchanting and Super Forging ignores enchantment compatibility";
         static final String splitEnchantmentRespectLevelExtension = "If Enchantment splitting respects over-capped level";